Refactor home page structure to use module layout
